+++ packages/ng/demo/app/api/api-select/custom/custom-select-feeder.ts
@@ -8,15 +8,21 @@ import { Injectable } from '@angular/core';
 
 @Injectable()
 export class DemoCustomApiSelectFeeder extends AApiSelectFeederWithPaging<any>
-	implements IApiSelectFeederWithPaging<any> {
+implements IApiSelectFeederWithPaging<any> {
+	private totalCountItems = 0;
+	private totalItems = [];
 	constructor(protected _http: HttpClient) {
 		super();
 	}
+
 	getPagedItems(
 		clue: string,
 		pagingStart: number,
 		pagingStep: number,
 	): Observable<any[]> {
+		if (pagingStart === 0 ) {
+			this.totalItems = [];
+		}
 		const values = [];
 		for (let i = pagingStart; i < pagingStart + pagingStep; i++) {
 			values.push({
@@ -24,6 +30,7 @@ export class DemoCustomApiSelectFeeder extends AApiSelectFeederWithPaging<any>
 				label: `Item ${clue} ${i}`,
 			});
 		}
+		this.totalCountItems += pagingStep;
 
 		return this._http
 			.get(
@@ -31,7 +38,11 @@ export class DemoCustomApiSelectFeeder extends AApiSelectFeederWithPaging<any>
 					JSON.stringify(values),
 				)}`,
 			)
-			.map(response => JSON.parse(decodeURIComponent((<any>response).values)));
+			.map(response => {
+				const items = JSON.parse(decodeURIComponent((<any>response).values));
+				this.totalItems.push(...items);
+				return items;
+			});
 	}
 
 	getPagingStep(): number {
@@ -40,5 +51,13 @@ export class DemoCustomApiSelectFeeder extends AApiSelectFeederWithPaging<any>
 
 	textValue(item: any): string {
 		return item.label;
+	}
+
+	length(): number {
+		return this.totalCountItems;
+	}
+
+	getAllEntities(): Observable<any[]> {
+		return Observable.of(this.totalItems);
 	}
 }