+++ packages/ng/src/app/user/select/user-select.int.ts
@@ -1,0 +1,8 @@
+import { LuSelectIntl } from '../../../../src/app/select/';
+
+export function getOverrideLuSelectIntl(): LuSelectIntl {
+	const luSelectIntl = new LuSelectIntl();
+	luSelectIntl.allTypeLabel = 'the users';
+	luSelectIntl.typeLabel = 'users';
+	return luSelectIntl;
+}